Moved expression lib.
