+++ tests/_internal/functools/test_retriable.py
@@ -6,7 +6,10 @@ from functools import reduce
 
 import pytest
 
-from faith._internal.functools.retriable import MaxAttemptsExceededError, RetryFunctionWrapper
+from faith._internal.functools.retriable import (
+    MaxAttemptsExceededError,
+    RetryFunctionWrapper,
+)
 
 
 class FailingFunction:
