+++ tests/_internal/parsing/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2025 Cisco Systems, Inc. and its affiliates
+#
+# SPDX-License-Identifier: Apache-2.0
+
+__all__: list[str] = []
